Changed NetworkConfig data model
The old NetworkConfig depended from VirtualMachine. Therefore, each
VirtualMachine had a network configuration that included an IPv4, a
hostname, and an optional IPv6. This may cause problems when we delete
and recreate a secondary VM because we may end up with different IPs
and hostnames associated to the secondary VM, affecting to the testing
DNS. With the change made here, now the secondary VM of a site will
always have the same private IPv4 address and hostname even if we
delete it and recreate it.