Improved way to clone VMs
We had to deal with the problem that before cloning VMs we had to
previously delete the target if it previously existed. Both tasks were
async therefore we needed a way to put the clone task after the
deletion process has completed or it will fail. This is corrected now.