Added 2 funcs to the CheckBannedUsers middleware
Check that the user is in the jackdaw local cache: It is very unlikely
that the user can enter using raven without being in jackdaw but it may
be possible that our local cache is outdated because we update it only
once a day, therefore it may be possible.
Check that the user is active. When the user no longer appears in the
jackdaw database, the user is deactivated instead of being deleted.
This may change in the future.