Move initial SSH authorized keys for VMs from preseed file to Python.
It seems that it's only possible to specify one preseed/late_command,
so the callback specified by apimws.platforms.install_vm was overriding
the one in the base preseed file, with the consequence that the
standard SSH authorized keys were not getting installed.  The entire
command is now constructed in apimws.platforms.install_vm, which
means that SSH keys are now installed properly.