HostNetworkConfig modified
Added IPv4, IPv4_netmask, IPv4_gateway, and type.
Type specifies if it is a service network configuration or a host
network configuration. IPv4 parameters or IPv6 can be null as this
allows more flexibility in the possible Network Configurations (IPv6
only, IPv4 only, etc). hostname parameter is no longer called hostname
but name, as it can be a FQDN or other.