add docker-compose configuration to boostrap development environment
Make use of docker-compose to create a development environment on a
modern machine. This commit tries to replicate the environment we
currently run in production including the custom Debian repositories we
use, their keys and the custom Python packages.

The docker-compose configuration makes use of Postgres within a
container as the database. The container uses apache to serve the
application.

Documentation is added under doc/ to give developers an idea of how to
get started with the development site.

The Docker container is based on the debian Docker image with Python
dependencies installed via pip rather than APT. This lets us drop a
requirement to use third-party repositories.

Update developer-getting-started.md to cover the changes.