code review changes
Based on the information in the MWS DNS API docs, we don't need to
make sure the SSHFP keys exist, as deleting a nonexistent key will
not throw.
This commit also renames find_sshfp to get_sshfp to be in line with
the other utility functions and corrects the typo that made it not
work.